# [Static Runtime] Enable manage_output_tensors for ptvsc2_predictor_bench