Summary: This change add benchmark support for manage_output_tensors from ptvsc2_predictor_bench.

Test Plan:
- Confirmed that Static Runtime can run inline_cvr's local, local_ro, remote_ro by ptvsc2_predictor_bench.

- Confirmed that enabling `manage_output_tensors` accelerated the performance of Static Runtime as follows (note that there's an observed regression on remote_ro, and I observed swings too. The enclosed numbers are the median among 3 runs):
-- inline_cvr/local: 6.75441ms -> 6.40759ms (-5.1%)
-- inline_cvr/local_ro: 2.0251ms -> 1.65929ms (-18%)
-- inline_cvr/remote_ro: 224.301ms ->  225.322ms (+0.4%)

- Confirmed that enabling `manage_output_tensors` consistently lets the runtime manage output tensors:
-- inline_cvr/local: unmanaged values from 3239 to 2272
-- inline_cvr/local_ro:  unmanaged values from 2966 to 1836
-- inline_cvr/remote_ro: unmanaged values from 28 to 0

- Check out the raw log output from `ptvsc2_predictor_bench` with `managed_output_tensors` : P460137877 (generated by script P460137907).

Differential Revision: D31149323